LayoutTests/http/tests/security/canvas-remote-read-svg-image-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
This tests that drawing a SVG image to a canvas taints the canvas

PASS: getImageData failed. Canvas tainted.
  
